added the buttons, with functions
