+++ Assets/Prefabs/MusicButton.prefab.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 4250195802ebc4f2fb076f0a67ab2e2b
-NativeFormatImporter:
+guid: a4f9cee9dc5fd4be5a00a4e9174c9883
+PrefabImporter:
   externalObjects: {}
-  mainObjectFileID: 2100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
