+++ Assets/Scripts/MusicPlayer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d520d10ced35647039ac21cb5fb79991
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
