+++ Assets/Scripts/SoundPlayer.cs
@@ -1,0 +1,21 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class SoundPlayer : MonoBehaviour
+{
+
+
+
+
+
+    bool audioOn = true;
+
+
+    public void Toggle()
+    {
+        audioOn = !audioOn;
+    }
+    
+     
+}
