+++ Assets/Scripts/SoundPlayer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f60d3a485f7bb4ce1b2b57e87cc8c398
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
